requirements.txt
django
django-filter
django-guardian
graphene[django]
graphql-django-view
django-graphiql